-Wall fixes
git-svn-id: https://svn.r-project.org/R/branches/R-0-65-patches@5877 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/devPS.c
@@ -252,8 +252,8 @@ static int KeyType(char *s)
 
 static char *SkipToNextItem(char *p)
 {
-    while (!isspace(*p)) p++;
-    while (isspace(*p)) p++;
+    while (!isspace((int)*p)) p++;
+    while (isspace((int)*p)) p++;
     return p;
 }
 
@@ -261,7 +261,7 @@ static char *SkipToNextKey(char *p)
 {
     while (*p != ';') p++;
     p++;
-    while (isspace(*p)) p++;
+    while (isspace((int)*p)) p++;
     return p;
 }
